service

+++ src/main/java/HSF301_Assignment_Spring_MVC/repositories/CarProducerRepository.java
@@ -3,5 +3,5 @@ package HSF301_Assignment_Spring_MVC.repositories;
 import HSF301_Assignment_Spring_MVC.pojos.CarProducer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarProducerRepository extends JpaRepository<CarProducer, Long> {
+public interface CarProducerRepository extends JpaRepository<CarProducer, Integer> {
 }
