+++ src/main/java/HSF301_Assignment_Spring_MVC/repositories/CarRepository.java
@@ -3,5 +3,7 @@ package HSF301_Assignment_Spring_MVC.repositories;
 import HSF301_Assignment_Spring_MVC.pojos.Car;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarRepository extends JpaRepository<Car, Long> {}
+public interface CarRepository extends JpaRepository<Car, Integer> {
+
+}
